Pipeline refactor: Introducing Pipeline werkzeug
- Moved tools module into pipeline package
- Moved `cwl_file_path_as_string_to_dict` into tools
- Deleted un-use metadata_globals.py
