+++ data_processors/pipeline/tests/test_tools.py
@@ -1,0 +1,51 @@
+import json
+
+from data_processors.pipeline import tools
+from data_processors.pipeline.tests.case import PipelineUnitTestCase, logger
+
+
+class ToolsUnitTests(PipelineUnitTestCase):
+
+    def test_parse_bcl_convert_output(self):
+        """
+        python manage.py test data_processors.pipeline.tests.test_tools.ToolsUnitTests.test_parse_bcl_convert_output
+        """
+
+        result = tools.parse_bcl_convert_output(json.dumps({
+            "main/fastq_list_rows": [{'rgid': "main/fastq_list_rows"}],
+            "fastq_list_rows": [{'rgid': "YOU_SHOULD_NOT_SEE_THIS"}]
+        }))
+
+        logger.info("-" * 32)
+        logger.info(f"parse_bcl_convert_output: {json.dumps(result)}")
+
+        self.assertEqual(result[0]['rgid'], "main/fastq_list_rows")
+
+    def test_parse_bcl_convert_output_alt(self):
+        """
+        python manage.py test data_processors.pipeline.tests.test_tools.ToolsUnitTests.test_parse_bcl_convert_output_alt
+        """
+
+        result = tools.parse_bcl_convert_output(json.dumps({
+            "fastq_list_rows2": [{'rgid': "YOU_SHOULD_NOT_SEE_THIS"}],
+            "fastq_list_rows": [{'rgid': "fastq_list_rows"}]
+        }))
+
+        logger.info("-" * 32)
+        logger.info(f"parse_bcl_convert_output alt: {json.dumps(result)}")
+
+        self.assertEqual(result[0]['rgid'], "fastq_list_rows")
+
+    def test_parse_bcl_convert_output_error(self):
+        """
+        python manage.py test data_processors.pipeline.tests.test_tools.ToolsUnitTests.test_parse_bcl_convert_output_error
+        """
+
+        try:
+            tools.parse_bcl_convert_output(json.dumps({
+                "fastq_list_rows/main": [{'rgid': "YOU_SHOULD_NOT_SEE_THIS"}],
+                "fastq_list_rowz": [{'rgid': "YOU_SHOULD_NOT_SEE_THIS_TOO"}]
+            }))
+        except Exception as e:
+            logger.exception(f"THIS ERROR EXCEPTION IS INTENTIONAL FOR TEST. NOT ACTUAL ERROR. \n{e}")
+        self.assertRaises(json.JSONDecodeError)
