--- utils/metadata_globals.py
@@ -1,8 +1,0 @@
-#!/usr/bin/env python
-
-"""
-Metadata constants
-"""
-
-
-TSO_TYPES = ["ctDNA", "TSO-DNA"]
